hair_shop/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.2
django-cleanup==9.0.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-unfold==0.89.0
et_xmlfile==2.0.0
idna==3.11
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pillow==12.1.1
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
yt-dlp==2026.3.17
